+++ src/isSymbol.js
@@ -19,11 +19,11 @@ import { type, curryN } from 'ramda';
  * RA.isSymbol(undefined); //=> false
  * RA.isSymbol(null); //=> false
  */
-const isSymbol = curryN(1, (val) => {
-  return (
+const isSymbol = curryN(
+  1,
+  (val) =>
     typeof val === 'symbol' ||
     (typeof val === 'object' && type(val) === 'Symbol')
-  );
-});
+);
 
 export default isSymbol;